fix(children): avatar_id is a semantic string end to end
The onboarding picker sends avatar ids like "cat"/"dolphin", every API
response already returns avatar_id as a string, and the frontend art
lookup is keyed by name — but the domain/db typed it u8/SMALLINT, so
real onboarding data could never validate. avatar_id is now TEXT/String
throughout (reversible migration carries numeric legacy rows over as
digit strings, which the UI already tolerates via its emoji fallback).
Domain validation: 1-32 chars, alnum/_/-. .sqlx cache regenerated for
the child_repo query! macros; test fixtures updated.

+++ backend/crates/infra/src/child_repo.rs
@@ -45,7 +45,7 @@ impl ChildRepo for SqlxChildRepo {
             p.id,
             p.parent_account_id,
             p.nickname,
-            p.avatar_id as i16,
+            p.avatar_id,
             p.birth_year as i16,
             p.created_at,
         )
@@ -69,7 +69,7 @@ impl ChildRepo for SqlxChildRepo {
             id: r.id,
             parent_account_id: r.parent_account_id,
             nickname: r.nickname,
-            avatar_id: r.avatar_id as u8,
+            avatar_id: r.avatar_id,
             birth_year: r.birth_year as u16,
             created_at: r.created_at,
         }))
@@ -91,7 +91,7 @@ impl ChildRepo for SqlxChildRepo {
                 id: r.id,
                 parent_account_id: r.parent_account_id,
                 nickname: r.nickname,
-                avatar_id: r.avatar_id as u8,
+                avatar_id: r.avatar_id,
                 birth_year: r.birth_year as u16,
                 created_at: r.created_at,
             })
